Crate sts

Symbol Table Stack Module

This module is responsible for providing the representation and generation of a symbol table stack for later stages of the compiler.

Structure

The sts module is organized into submodules that handle different aspects of the symbol table stack:

core: Defines the symbol table stack structure and drives the generation process.

statement: Handles stack generation for statements and operations within blocks.

block: Handles stack generation for code blocks.

Modules

core

Definitions of the symbol table stack structure and core of the symbol table stack generation process This file contains the core definitions and functionalities of the SymbolTableStack (STS).