Module sts::fullscreen_effect

Modules

descriptors 🔒

fullscreen_effect_builder 🔒

Tools for manipulating the type state of FullscreenEffectBuilder.

pipeline 🔒

Structs

FrameData

FullscreenEffect

A fullscreen effect is a reloadable fragment shader that is provided with updated data for each frame.

FullscreenEffectBuilder

Use builder syntax to set the inputs and finish with build().