Module sts::fullscreen_effect::fullscreen_effect_builder

Tools for manipulating the type state of FullscreenEffectBuilder.

See the detailed guide that describes how all the pieces here fit together.

Modules

sealed 🔒

Structs

Empty

Represents a State that has IsUnset implemented for all members.

SetCtx

Represents a State that has IsSet implemented for State::Ctx.

SetEffectShader

Represents a State that has IsSet implemented for State::EffectShader.

SetFramesInFlight

Represents a State that has IsSet implemented for State::FramesInFlight.

SetRenderPass

Represents a State that has IsSet implemented for State::RenderPass.

SetTextureAtlasLayout

Represents a State that has IsSet implemented for State::TextureAtlasLayout.

Traits

IsComplete

Marker trait that indicates that all required members are set.

IsSet

Marker trait that indicates that the member is set, i.e. at least one of its setters was called.

IsUnset

Marker trait that indicates that the member is unset, i.e. none of its setters was called.

State

Builder’s type state specifies if members are set or not (unset).