Trait sts::fullscreen_effect::fullscreen_effect_builder::State

pub trait State: Sized {
    type Ctx;
    type FramesInFlight;
    type TextureAtlasLayout;
    type RenderPass;
    type EffectShader;
}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type Ctx

Type state of the member ctx.

It can implement either IsSet or IsUnset

type FramesInFlight

Type state of the member frames_in_flight.

It can implement either IsSet or IsUnset

type TextureAtlasLayout

Type state of the member texture_atlas_layout.

It can implement either IsSet or IsUnset

type RenderPass

Type state of the member render_pass.

It can implement either IsSet or IsUnset

type EffectShader

Type state of the member effect_shader.

It can implement either IsSet or IsUnset

Object Safety

This trait is not object safe.