Module sts::graphics::vulkan::raii

RAII wrappers for Vulkan objects.

Wrappers do not track dependencies. The application is responsible for dropping Vulkan objects in the correct order and synchronizing to prevent GPU inconsistencies.

Structs

Buffer

RAII wrapper that destroys itself when Dropped.

CommandPool

RAII wrapper that destroys itself when Dropped.

DebugUtils

RAII wrapper that destroys itself when Dropped.

DescriptorPool

RAII wrapper that destroys itself when Dropped.

DescriptorSetLayout

RAII wrapper that destroys itself when Dropped.

Device

A RAII wrapper for the Vulkan Logical Device.

DeviceMemory

RAII wrapper that destroys itself when Dropped.

Fence

RAII wrapper that destroys itself when Dropped.

Framebuffer

RAII wrapper that destroys itself when Dropped.

Image

RAII wrapper that destroys itself when Dropped.

ImageView

RAII wrapper that destroys itself when Dropped.

Instance

A RAII wrapper for the ash library entry and instance that destroys itself when dropped.

Pipeline

RAII wrapper that destroys itself when Dropped.

PipelineLayout

RAII wrapper that destroys itself when Dropped.

RenderPass

RAII wrapper that destroys itself when Dropped.

Semaphore

RAII wrapper that destroys itself when Dropped.

ShaderModule

RAII wrapper that destroys itself when Dropped.

Surface

RAII wrapper that destroys itself when Dropped.

Swapchain

RAII wrapper that destroys itself when Dropped.